refactor(hax-lib): inline the small proc-macro implementations next to their documentation in `lib.rs`.
